==> 02_calculate_delta.ipynb <==
# Project 01 - Data Processing

## Step 5 - Calulate Additional Data

import datetime
import numpy as np
import pandas as pd
pd.__version__

BASE_PATH = "./data/jh_covid19/csse_covid_19_data/csse_covid_19_daily_reports/"
BASE_PATH_WB = "./data/world_bank/"
BASE_PATH_OUT = "./data/output/"

df_src = pd.read_csv(BASE_PATH_OUT + 'covid_data.csv')
df_src['date']= pd.to_datetime(df_src['date'])
df_src

### 5.1 Calculate Deltas

list_codes = df_src['Short Code'].unique()
list_codes

dfs_by_regions = {}

def get_fatal_rate(confirmed, death):
    if confirmed == 0:
        return 0.
    else:
        return float(death) / float(confirmed)

    
def get_zero_or_result(a, b):
    return 0 if a < b else a - b
    
    
for i in range(len(list_codes)):
    print('processing: ' + list_codes[i])
    df_current_by_region = df_src[df_src['Short Code'] == list_codes[i]]
    df_current_by_region = df_current_by_region.sort_values(by=['date'])
    
    list_confirmed = df_current_by_region['Confirmed'].tolist()
    list_death = df_current_by_region['Deaths'].tolist()
    
    list_delta_confirmed = []
    list_delta_death = []
    list_fatality_rate = []
    
    for j in range(len(list_confirmed)):
        list_fatality_rate.append(get_fatal_rate(list_confirmed[j], list_death[j]))
        if j == 0:
            list_delta_confirmed.append(list_confirmed[j])
            list_delta_death.append(list_death[j])
        else:
            list_delta_confirmed.append(get_zero_or_result(list_confirmed[j], list_confirmed[j - 1]))
            list_delta_death.append(get_zero_or_result(list_death[j], list_death[j - 1]))
            
    df_current_by_region['New Confirmed'] = list_delta_confirmed
    df_current_by_region['New Deaths'] = list_delta_death
    df_current_by_region['Fatality Ratio'] = list_fatality_rate
    
    dfs_by_regions[list_codes[i]] = df_current_by_region

dfs_by_regions['CHN']

df_current = dfs_by_regions['GBR']

df_current[df_current['date'] == datetime.datetime.strptime('2022-05-25', '%Y-%m-%d')]

df_current = dfs_by_regions['GBR']

df_current[df_current['date'] == datetime.datetime.strptime('2022-05-26', '%Y-%m-%d')]

### 5.2 Export

dfs_merged_regions = pd.concat(list(dfs_by_regions.values())).reset_index(drop=True)

dfs_merged_regions.columns = ['Country Region Name', 'Confirmed', 'Deaths', 'Date', 'Country Code', 'New Confirmed', 'New Deaths', 'Fatality Ratio']

dfs_merged_regions.head()

dfs_merged_regions = dfs_merged_regions.sort_values(by=["Country Code", "Date"]).reset_index(drop=True)

dfs_merged_regions.to_csv(BASE_PATH_OUT + "covid_data_v3.csv", date_format='%m/%d/%Y', index=False)

# sanity check
dfs_merged_regions[dfs_merged_regions['Country Code'] == 'CHN']
