==> 03_join_tables.ipynb <==
# Project 01 - Data Processing

## Step 6 - Manually Join Tables

import datetime
import numpy as np
import pandas as pd
pd.__version__

BASE_PATH = "./data/jh_covid19/csse_covid_19_data/csse_covid_19_daily_reports/"
BASE_PATH_WB = "./data/world_bank/"
BASE_PATH_OUT = "./data/output/"

df_src = pd.read_csv(BASE_PATH_OUT + 'covid_data_v3.csv')
df_src['Date']= pd.to_datetime(df_src['Date'])
df_src

### 6.1 Generate Formal Country Name Map

df_countries = pd.read_csv(BASE_PATH_WB + 'WDICountry.csv')
df_countries.head()

country_map_by_code = {}
country_code_list = df_countries['Country Code'].unique()
country_code_list

df_countries[df_countries['Country Code'] == 'GBR'].iloc[0]['Short Name']

for i in range(len(country_code_list)):
    current_country_code = country_code_list[i]
    current_country_name = df_countries[df_countries['Country Code'] == current_country_code].iloc[0]['Short Name']
    country_map_by_code[current_country_code] = current_country_name


country_map_by_code['GBR']

### 6.2 Generate Life Expectancy Map

df_wdi = pd.read_csv(BASE_PATH_WB + 'WDIData.csv')
df_wdi.head()

df_wdi[(df_wdi['Country Code'] == 'GBR') & (df_wdi['Indicator Code'] == 'SP.DYN.LE00.IN')].iloc[0]['2020']


life_expectancy_map = {}
for i in range(len(country_code_list)):
    current_country_code = country_code_list[i]
    current_country_name = df_wdi[(df_wdi['Country Code'] == current_country_code) & (df_wdi['Indicator Code'] == 'SP.DYN.LE00.IN')].iloc[0]['2020']
    life_expectancy_map[current_country_code] = current_country_name


life_expectancy_map['GBR']

### 6.3 Generate Total Population Map

df_wdi[(df_wdi['Country Code'] == 'GBR') & (df_wdi['Indicator Code'] == 'SP.POP.TOTL')].iloc[0]['2020']


total_population_map = {}
for i in range(len(country_code_list)):
    current_country_code = country_code_list[i]
    current_country_name = df_wdi[(df_wdi['Country Code'] == current_country_code) & (df_wdi['Indicator Code'] == 'SP.POP.TOTL')].iloc[0]['2020']
    total_population_map[current_country_code] = current_country_name


total_population_map['GBR']

### 6.4 Manually Join Tables

df_src['Country Name'] = df_src['Country Code'].map(country_map_by_code)
df_src['Life Expectancy'] = df_src['Country Code'].map(life_expectancy_map)
df_src['Total Population'] = df_src['Country Code'].map(total_population_map)
df_src

# validate
life_expectancy_map['AFG']

life_expectancy_map['ZWE']

total_population_map['AFG']

total_population_map['ZWE']

### 6.5 Build Final Table

df_output = df_src[['Country Code', 'Country Name', 'Date', 'Confirmed', 'Deaths', 'New Confirmed', 'New Deaths', 'Fatality Ratio', 'Life Expectancy', 'Total Population']]
df_output


df_output.to_csv(BASE_PATH_OUT + "covid_data_v4.csv", date_format='%m/%d/%Y', index=False)
