==> 04_data_processing_part2.ipynb <==
# Project 02 - Data Processing

## Step 4 - Calculate Weekly Average

import datetime
import numpy as np
import pandas as pd
pd.__version__

BASE_PATH = "./data/jh_covid19/csse_covid_19_data/csse_covid_19_daily_reports/"
BASE_PATH_WB = "./data/world_bank/"
BASE_PATH_WHO = "./data/who/"
BASE_PATH_OUT = "./data/output/"

df_src = pd.read_csv(BASE_PATH_OUT + 'covid_data_v4.csv')
df_src['Date']= pd.to_datetime(df_src['Date'])
df_src

df_src.dtypes

### 4.1 Examine Data

country_map_by_code = {}
country_code_list = df_src['Country Code'].unique()
country_code_list

start_date_list = []
skip_list = []
GLOBAL_START_DATE = datetime.datetime.strptime('2020-04-27', '%Y-%m-%d')

for i in range(len(country_code_list)):
    current_country = country_code_list[i]
    df_current = df_src[df_src['Country Code'] == current_country]
    start_date = df_current['Date'].min()
    
    if start_date > GLOBAL_START_DATE:
        print('skip: {}: {}'.format(current_country, start_date))
        skip_list.append(current_country)
        continue
    
    print('{}: {}'.format(current_country, start_date))
    start_date_list.append(start_date)
    
max(start_date_list)

df_countries = pd.read_csv(BASE_PATH_WB + 'WDICountry.csv')
df_countries.head()

for i in range(len(skip_list)):
    current_country = skip_list[i]
    df_result = df_countries[df_countries['Country Code'] == current_country]
    print('{} = {}'.format(current_country, df_result.iloc[0]['Short Name']))

country_code_list = [x for x in country_code_list if x not in skip_list]
np.array(country_code_list)

### 4.2 Preparation

date_list = []
monday_date = GLOBAL_START_DATE

while monday_date <= datetime.datetime.strptime('2022-05-30', '%Y-%m-%d'):
    date_list.append(monday_date)
    monday_date += datetime.timedelta(days = 7)
    
date_list = np.array(date_list)
date_list

df_test = df_src[(df_src['Date'] >= date_list[0]) & (df_src['Date'] < date_list[1]) & (df_src['Country Code'] == 'USA')]
df_test

df_test['New Confirmed'].sum()

len(df_test)

### 4.3 Generate Data

df_avg_by_region = {}

for i in range(len(country_code_list)):
    current_country = country_code_list[i]
#     current_country = 'USA'

    print('processing: {} ...'.format(current_country))
    df_current = pd.DataFrame(date_list[:-1], columns=['Date'])
    df_current['Country Code'] = current_country
    
    col_avg_new = []
    col_avg_death = []
    col_avg_fatality = []
    
    for j in range(len(date_list) - 1):
        current_start_date = date_list[j]
        current_end_date = date_list[j + 1]
        
        df_filtered = df_src[(df_src['Date'] >= current_start_date) & (df_src['Date'] < current_end_date) & (df_src['Country Code'] == current_country)]
        number_record = len(df_filtered)
        
        col_avg_new.append(df_filtered['New Confirmed'].sum() / float(number_record))
        col_avg_death.append(df_filtered['New Deaths'].sum() / float(number_record))
        col_avg_fatality.append(df_filtered['Fatality Ratio'].sum() / float(number_record))
        
    df_current['7 Day Avg New Confirmed'] = col_avg_new
    df_current['7 Day Avg New Deaths'] = col_avg_death
    df_current['7 Day Avg Fatality Ratio'] = col_avg_fatality
    
    df_avg_by_region[current_country] = df_current


df_avg_by_region['GBR']

### 4.4 Validation

df_test = df_src[
    (df_src['Date'] >= datetime.datetime.strptime('2022-05-16', '%Y-%m-%d')) & 
    (df_src['Date'] < datetime.datetime.strptime('2022-05-23', '%Y-%m-%d')) & 
    (df_src['Country Code'] == 'GBR')]
df_test

df_test['Fatality Ratio'].sum() / 7.

df_test['New Confirmed'].sum() / 7.

df_test['New Deaths'].sum() / 7.

### 4.5 Export

dfs_merged_regions = pd.concat(list(df_avg_by_region.values())).reset_index(drop=True)
dfs_merged_regions

dfs_merged_regions = dfs_merged_regions.sort_values(by=["Country Code", "Date"]).reset_index(drop=True)

dfs_merged_regions.to_csv(BASE_PATH_OUT + "covid_data_v5.csv", date_format='%m/%d/%Y', index=False)
