==> 05_additional_data.ipynb <==
# Project 02 - Data Processing

## Step 5 - Additional Data

import datetime
import numpy as np
import pandas as pd
pd.__version__

BASE_PATH = "./data/jh_covid19/csse_covid_19_data/csse_covid_19_daily_reports/"
BASE_PATH_WB = "./data/world_bank/"
BASE_PATH_WHO = "./data/who/"
BASE_PATH_OUT = "./data/output/"

df_countries = pd.read_csv(BASE_PATH_WB + 'WDICountry.csv')
df_countries.head()

df_statistic = pd.read_csv(BASE_PATH_WB + 'WDIData.csv')
df_statistic.head()

df_data = pd.read_csv(BASE_PATH_OUT + 'covid_data_v5.csv')
df_data['Date']= pd.to_datetime(df_data['Date'])
df_data

country_map_by_code = {}
country_code_list = df_data['Country Code'].unique()
country_code_list

len(df_statistic['Indicator Name'].unique())

selected_years = list(range(1980, 2021, 5))
NUM_OF_YEARS = len(selected_years)
selected_years

### 5.1 Prepare Columns

#### 5.1.1 Country Name, Country Code, Years

col_country_names = []

for i in range(len(country_code_list)):
    current_country = country_code_list[i]
    df_filtered = df_countries[df_countries['Country Code'] == current_country]
    
    for j in range(len(selected_years)):
        col_country_names.append(df_filtered.iloc[0]['Short Name'])
    
print(len(col_country_names))

col_country_codes = []

for i in range(len(country_code_list)):
    current_country = country_code_list[i]
    
    for j in range(len(selected_years)):
        col_country_codes.append(current_country)
    
print(len(col_country_codes))

col_years = []

for i in range(len(country_code_list)):
    for j in range(len(selected_years)):
        col_years.append(selected_years[j])
    
print(len(col_years))
# col_years

#### 5.1.2 Life expectancy at birth, total (years)

__SP.DYN.LE00.IN__

df_filtered = df_statistic[(df_statistic['Country Code'] == 'USA') & (df_statistic['Indicator Code'] == 'SP.DYN.LE00.IN')]
print(df_filtered)
for j in range(len(selected_years)):
    print(df_filtered.iloc[0][str(selected_years[j])])


col_sp_dyn_le00_in = []
current_indicator = 'SP.DYN.LE00.IN'

for i in range(len(country_code_list)):
    current_country = country_code_list[i]
#     print('processing: {} ...'.format(current_country))
    df_filtered = df_statistic[(df_statistic['Country Code'] == current_country) & (df_statistic['Indicator Code'] == current_indicator)]
#     print(df_filtered)
    
    for j in range(len(selected_years)):
        col_sp_dyn_le00_in.append(df_filtered.iloc[0][str(selected_years[j])])
#         break

#     break
    
# col_sp_dyn_le00_in


#### 5.1.3 Population, total

__SP.POP.TOTL__

col_sp_pop_totl = []
current_indicator = 'SP.POP.TOTL'

for i in range(len(country_code_list)):
    current_country = country_code_list[i]
#     print('processing: {} ...'.format(current_country))
    df_filtered = df_statistic[(df_statistic['Country Code'] == current_country) & (df_statistic['Indicator Code'] == current_indicator)]
#     print(df_filtered)
    
    for j in range(len(selected_years)):
        col_sp_pop_totl.append(df_filtered.iloc[0][str(selected_years[j])])


#### 5.1.4 Human Capital Index (HCI) (scale 0-1)

__HD.HCI.OVRL__

col_hd_hci_ovrl = []
current_indicator = 'HD.HCI.OVRL'

for i in range(len(country_code_list)):
    current_country = country_code_list[i]
#     print('processing: {} ...'.format(current_country))
    df_filtered = df_statistic[(df_statistic['Country Code'] == current_country) & (df_statistic['Indicator Code'] == current_indicator)]
#     print(df_filtered)
    
    for j in range(len(selected_years)):
        col_hd_hci_ovrl.append(df_filtered.iloc[0][str(selected_years[j])])


#### 5.1.5 GDP per capita (current US$)

__NY.GDP.PCAP.CD__

col_ny_gdp_mktp_cd = []
current_indicator = 'NY.GDP.PCAP.CD'

for i in range(len(country_code_list)):
    current_country = country_code_list[i]
#     print('processing: {} ...'.format(current_country))
    df_filtered = df_statistic[(df_statistic['Country Code'] == current_country) & (df_statistic['Indicator Code'] == current_indicator)]
#     print(df_filtered)
    
    for j in range(len(selected_years)):
        col_ny_gdp_mktp_cd.append(df_filtered.iloc[0][str(selected_years[j])])


#### 5.1.6 GDP growth (annual %)

__NY.GDP.MKTP.KD.ZG__

col_ny_gdp_mktp_kd_zg = []
current_indicator = 'NY.GDP.MKTP.KD.ZG'

for i in range(len(country_code_list)):
    current_country = country_code_list[i]
#     print('processing: {} ...'.format(current_country))
    df_filtered = df_statistic[(df_statistic['Country Code'] == current_country) & (df_statistic['Indicator Code'] == current_indicator)]
#     print(df_filtered)
    
    for j in range(len(selected_years)):
        col_ny_gdp_mktp_kd_zg.append(df_filtered.iloc[0][str(selected_years[j])])


### 5.2 Generate Table

df_result = pd.DataFrame(col_country_names, columns=['Country Name'])
df_result

df_result['Country Code'] = col_country_codes
df_result['Year'] = col_years

df_result['SP.DYN.LE00.IN'] = col_sp_dyn_le00_in
df_result['SP.POP.TOTL'] = col_sp_pop_totl
# df_result['HD.HCI.OVRL'] = col_hd_hci_ovrl
df_result['NY.GDP.PCAP.CD'] = col_ny_gdp_mktp_cd
df_result['NY.GDP.MKTP.KD.ZG'] = col_ny_gdp_mktp_kd_zg

df_result = df_result.sort_values(by=["Country Code", "Year"]).reset_index(drop=True)

df_result.to_csv(BASE_PATH_OUT + "wdi_data_v1.csv", date_format='%m/%d/%Y', index=False)

df_result

### 5.3 Validation

df_val = pd.read_csv(BASE_PATH_OUT + 'wdi_data_v1.csv')
df_val

df_val[df_val['Country Code'] == 'USA']

df_filtered = df_statistic[(df_statistic['Country Code'] == 'GBR') & (df_statistic['Indicator Code'] == 'NY.GDP.PCAP.CD')]
# print(df_filtered)
for j in range(len(selected_years)):
    print(df_filtered.iloc[0][str(selected_years[j])])


# !!! TEST
df_val['SP.POP.TOTL'].max()

df_val[df_val['Country Code'] == 'USA']['SP.DYN.LE00.IN'].max()
