==> 06_country_data.ipynb <==
# Project 02 - Data Processing

## Step 6 - Generate Country Data

import datetime
import numpy as np
import pandas as pd
pd.__version__

BASE_PATH = "./data/jh_covid19/csse_covid_19_data/csse_covid_19_daily_reports/"
BASE_PATH_WB = "./data/world_bank/"
BASE_PATH_WHO = "./data/who/"
BASE_PATH_OUT = "./data/output/"

df_countries = pd.read_csv(BASE_PATH_WB + 'WDICountry.csv')
df_countries.head()

df_statistic = pd.read_csv(BASE_PATH_WB + 'WDIData.csv')
df_statistic.head()

df_vaccination = pd.read_csv(BASE_PATH_WHO + 'vaccination-data.csv')
df_vaccination.head()

df_data = pd.read_csv(BASE_PATH_OUT + 'covid_data_v5.csv')
df_data['Date']= pd.to_datetime(df_data['Date'])
df_data

df_covid = pd.read_csv(BASE_PATH_OUT + 'covid_data_v4.csv')
df_covid['Date']= pd.to_datetime(df_covid['Date'])
df_covid.head()

country_map_by_code = {}
country_code_list = df_data['Country Code'].unique()
country_code_list

df_result = pd.DataFrame(country_code_list, columns=['Country Code'])
df_result

col_country_names = []

for i in range(len(country_code_list)):
    current_value = df_countries[df_countries['Country Code'] == country_code_list[i]]
    col_country_names.append(current_value.iloc[0]['Short Name'])
    
# np.array(col_country_names)
len(col_country_names)

col_fatality_rate = []

for i in range(len(country_code_list)):
    current_value = df_data[(df_data['Date'] == datetime.datetime.strptime('2022-05-23', '%Y-%m-%d')) & (df_data['Country Code'] == country_code_list[i])]
    col_fatality_rate.append(current_value.iloc[0]['7 Day Avg Fatality Ratio'])
    
# np.array(col_fatality_rate)
len(col_fatality_rate)

col_confirmed_cases = []

for i in range(len(country_code_list)):
    current_value = df_covid[(df_covid['Date'] == datetime.datetime.strptime('2022-05-30', '%Y-%m-%d')) & (df_covid['Country Code'] == country_code_list[i])]
    col_confirmed_cases.append(current_value.iloc[0]['Confirmed'])
    
# np.array(col_confirmed_cases)
len(col_confirmed_cases)

col_first_vaccinated = []

for i in range(len(country_code_list)):
    current_value = df_vaccination[df_vaccination['ISO3'] == country_code_list[i]]
    col_first_vaccinated.append(current_value.iloc[0]['FIRST_VACCINE_DATE'])
    
# np.array(col_first_vaccinated)
len(col_first_vaccinated)

col_vaccinated_1shot_per100 = []

for i in range(len(country_code_list)):
    current_value = df_vaccination[df_vaccination['ISO3'] == country_code_list[i]]
    col_vaccinated_1shot_per100.append(current_value.iloc[0]['TOTAL_VACCINATIONS_PER100'])
    
# np.array(col_vaccinated_1shot_per100)
len(col_vaccinated_1shot_per100)

col_vaccinated_2shot_per100 = []

for i in range(len(country_code_list)):
    current_value = df_vaccination[df_vaccination['ISO3'] == country_code_list[i]]
    col_vaccinated_2shot_per100.append(current_value.iloc[0]['PERSONS_VACCINATED_1PLUS_DOSE_PER100'])
    
# np.array(col_vaccinated_2shot_per100)
len(col_vaccinated_2shot_per100)

col_vaccinated_3shot_per100 = []

for i in range(len(country_code_list)):
    current_value = df_vaccination[df_vaccination['ISO3'] == country_code_list[i]]
    col_vaccinated_3shot_per100.append(current_value.iloc[0]['PERSONS_FULLY_VACCINATED_PER100'])
    
# np.array(col_vaccinated_3shot_per100)
len(col_vaccinated_3shot_per100)

df_result['Country Name'] = col_country_names
df_result['Total Confirmed'] = col_confirmed_cases
df_result['Fatality Rate'] = col_fatality_rate
df_result['First Vaccination Date'] = col_first_vaccinated
df_result['1 Shot Per 100'] = col_vaccinated_1shot_per100
df_result['2 Shots Per 100'] = col_vaccinated_2shot_per100
df_result['3 Shots Per 100'] = col_vaccinated_3shot_per100

df_result

df_vaccination[df_vaccination['ISO3'] == country_code_list[i]]

df_vaccination[df_vaccination['ISO3'] == 'ARE']

import math

df_result[df_result['Fatality Rate'].isnull()]

df_result = df_result.sort_values(by=["Country Code"]).reset_index(drop=True)

df_result.to_csv(BASE_PATH_OUT + "country_data_v2.csv", date_format='%m/%d/%Y', index=False)

df_val = pd.read_csv(BASE_PATH_OUT + 'country_data_v2.csv')
df_val
